Fix project to have files in proper spots.
Currently, the master branch in thell3oss's repo is a bit screwy
because the main classes are not in src/. This commit fixes that.
