+++ .classpath
@@ -2,5 +2,6 @@
 <classpath>
 	<classpathentry kind="src" path="src"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
+	<classpathentry kind="lib" path="/Users/kofskyfamilycomputer/git/integrator/src/jep-2.23.jar"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
